requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.4.4
packaging==23.1
pluggy==1.3.0
PyJWT==1.7.1
pytest==7.4.1
pytz==2023.3
PyYAML==6.0.1
shortuuid==1.0.11
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
zipp==3.16.2
